Moved graph to a modal in node info page
